+++ packages/common/src/common/domain/model/AppError.ts
@@ -1,0 +1,10 @@
+import { AppErrorType } from './AppErrorType';
+
+export class AppError extends Error {
+  constructor(
+    message: string,
+    public readonly type: AppErrorType = AppErrorType.Unknown,
+  ) {
+    super(message);
+  }
+}