+++ packages/common/src/index.ts
@@ -1,9 +1,13 @@
 import { EntityCreationAdapter } from './common/domain/adapter/EntityCreationAdapter';
 import { EntityFindAdapter } from './common/domain/adapter/EntityFindAdapter';
 import { Converter } from './common/domain/Converter';
+import { AppError } from './common/domain/model/AppError';
+import { AppErrorType } from './common/domain/model/AppErrorType';
 import { Either } from './common/domain/model/Either';
 import { EitherEither } from './common/domain/model/EitherEither';
 import { ValueEither } from './common/domain/model/ValueEither';
 import { ValueOrErrors } from './common/domain/model/ValueOrErrors';
 
 export type { Converter, Either, EitherEither, EntityCreationAdapter, EntityFindAdapter, ValueEither, ValueOrErrors };
+
+export { AppError, AppErrorType };